pyproject.toml
[project]
name = "auto-archive-link"
version = "0.1.0"
description = "Discord bot that adds archive links"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "dotenv>=0.9.9",
]
